# app/static/swagger.yaml
swagger: '2.0'
info:
  title: 'Mechanic Shop API'
  description: 'API for a fictitious Mechanic Shop, built entirely from scratch.'
  version: '1.0.0'
host: '127.0.0.1:5000'
schemes:
  - 'http'
consumes:
  - 'application/json'
produces:
  - 'application/json'
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

# ------------------------
paths:
...
  /inventory:
    post:
      tags: [Inventory]
      summary: Add inventory item
      description: Endpoint to add a new inventory item/part to the system.
      parameters:
        - in: body
          name: body
          description: Information required to create a new Inventory Item
          required: true
          schema:
            $ref: '#/definitions/CreateInventoryItemPayload'
      responses:
        201:
          description: Successfully added inventory item
          schema:
            $ref: '#/definitions/InventoryItemResponse'
          examples:
            application/json:
              id: 1
              name: Oil Filter
              price: 19.99
        400:
          description: Duplicate item
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: 'Item already exists in inventory.'

    get:
      tags: [Inventory]
      summary: Returns list of all inventory items
      description: Endpoint to retrieve a list of all inventory items.
      responses:
        200:
          description: Items retrieved
          schema:
            type: array
            items:
              $ref: '#/definitions/InventoryItemResponse'
            example:
              application/json:
                - id: 1
                  name: Oil Filter
                  price: 19.99
                - id: 2
                  name: Spark Plug
                  price: 5.49
  /inventory/{id}:
    get:
      tags: [Inventory]
      summary: Get inventory item by ID
      description: Retrieve detailed information about a specific inventory item using its unique ID.
      parameters:
        - in: path
          name: id
          description: Inventory item's unique ID
          required: true
          type: integer
      responses:
        200:
          description: Item found
          schema:
            $ref: '#/definitions/InventoryItemResponse'
          examples:
            application/json:
              id: 1
              name: Oil Filter
              price: 19.99
        404:
          description: Item not found
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: 'Inventory item not found.'

    put:
      tags: [Inventory]
      summary: Update inventory item details
      description: Update an existing inventory item by its unique ID.
      parameters:
        - in: path
          name: id
          description: Inventory item's unique ID
          required: true
          type: integer
        - in: body
          name: body
          description: Information required to update an Inventory Item
          required: true
          schema:
            $ref: '#/definitions/UpdateInventoryItemPayload'
      responses:
        200:
          description: Item  details updated
          schema:
            $ref: '#/definitions/InventoryItemResponse'
          examples:
            application/json:
              id: 1
              name: Oil Filter
              price: 21.99
        404:
          description: Item not found
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: 'Inventory item not found.'

    delete:
      tags: [Inventory]
      summary: Delete inventory item by ID
      description: Delete an inventory item using its unique ID.
      parameters:
        - in: path
          name: id
          description: Inventory item's unique ID
          required: true
          type: integer
      responses:
        200:
          description: Successfully deleted inventory item
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: 'Inventory item <id> deleted successfully'
        404:
          description: Item not found
          schema:
            type: object
            properties:
              message:
                type: string
            example:
              message: 'Inventory item not found.'

# ------------------------
definitions:
...
  CreateInventoryItemPayload:
    type: object
    properties:
      name:
        type: string
      price:
        type: number
    required: [name, price]
  UpdateInventoryItemPayload:
    $ref: '#/definitions/CreateInventoryItemPayload'
  InventoryItemResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number